telegram-bot/requirements.txt
python-telegram-bot==21.6
openai>=1.40.0
google-genai>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.4.0
